MCP_3D/Automation/desktop-bot/requirements.txt
lxml==5.3.2
python-dotenv==1.1.0
apscheduler==3.11.0
schedule==1.2.2
requests==2.32.3
flask==3.1.0
flask-cors==5.0.1
flask-limiter==3.12
gunicorn==23.0.0
pytest==8.3.5
werkzeug==3.1.3
termcolor==3.0.1
pyautogui==0.9.54
paramiko==3.5.1
pytesseract==0.3.13
pillow==11.2.1
opencv-python==4.11.0.86
psutil==7.0.0
numpy==2.2.4
pyperclip==1.9.0
pygetwindow; platform_system == "Windows"
pywinctl==0.4.1; platform_system == "Linux" or platform_system == "Darwin"
imaplib2==3.6
email-validator==2.2.0
spacy==3.8.5
transformers==4.51.3
torch==2.6.0
tensorflow; platform_system != "Windows" or python_version < "3.10"
tensorflow-cpu; platform_system == "Windows" and python_version >= "3.10"
SpeechRecognition==3.14.2
pyaudio==0.2.14
python-Levenshtein==0.27.1
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
scikit-learn==1.6.1
pytest-cov==6.1.1
pytest-mock==3.14.0
